Move edge types into a dedicated mod
